refactor(domain,view): 도메인 로직과 UI 로직을 각각 domain,view 폴더로 이동
