+++ src/App.js
@@ -1,3 +1,3 @@
-const RacingGame = require('./controller/RacingGame');
+const RacingGame = require('./domain/controller/RacingGame');
 
 new RacingGame().play();
